+++ backend/src/main/java/ch/web/web_shop/dto/cart/AddToCartDto.java
@@ -3,9 +3,9 @@ package ch.web.web_shop.dto.cart;
 import jakarta.validation.constraints.NotNull;
 
 public class AddToCartDto {
-    private Integer id;
+    private long id;
     @NotNull
-    private Integer productId;
+    private long productId;
     @NotNull
     private Integer quantity;
 
@@ -23,20 +23,20 @@ public class AddToCartDto {
                 ",";
     }
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
 
 
-    public Integer getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(Integer productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
